pyproject.toml
[project]
name = "babybuddy"
version = "0.1.0"
description = "Python-first baby tracker app"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "Django>=5.1,<5.3",
  "psycopg[binary]>=3.2,<4",
  "python-dotenv>=1.0,<2",
  "whitenoise>=6.7,<7",
  "gunicorn>=23.0,<24",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "babybuddy.settings.local"
python_files = ["test_*.py", "*_tests.py"]

[tool.ruff]
line-length = 100

target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
